+++ packages/Checkbox/tests/Checkbox.spec.js
@@ -5,7 +5,7 @@ import Checkbox from "../src/Checkbox";
 function renderComponent(props = {}) {
   const defaultProps = {
     a11yText: null,
-    checkedState: Checkbox.States.CHECKED,
+    checkedState: Checkbox.states.CHECKED,
     children: null,
     isDisabled: false,
     onChange: () => {},
@@ -21,12 +21,12 @@ describe("Checkbox", () => {
   });
 
   it("Renders checkedState is unchecked", () => {
-    const { getByRole } = renderComponent({ checkedState: Checkbox.States.UNCHECKED });
+    const { getByRole } = renderComponent({ checkedState: Checkbox.states.UNCHECKED });
     expect(getByRole("checkbox")).not.toHaveAttribute("checked");
   });
 
   it("Renders checkedState is indeterminate", () => {
-    const { getByRole } = renderComponent({ checkedState: Checkbox.States.INDETERMINATE });
+    const { getByRole } = renderComponent({ checkedState: Checkbox.states.INDETERMINATE });
     expect(getByRole("checkbox").indeterminate).toBe(true);
   });