+++ packages/Checkbox/CHANGELOG.md
@@ -5,4 +5,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.2.0] - 2020-01-14
 ### Added
-- Fixed broken checkboxStates import to use Checkbox.States instead [@oscarkwan](https://github.com/oscarkwan).
+- Fixed broken checkboxStates import to use Checkbox.states instead [@oscarkwan](https://github.com/oscarkwan).